+++ app/models/install.rb
@@ -1,5 +1,7 @@
 class Install < ActiveRecord::Base
 
+  include IdentifierObfuscation
+
   belongs_to :user
 
 end
